requirements.txt
google-genai>=1.33.0
Pillow>=11.0.0
flask>=3.1.0
